Zeming's fixes for Windoze
