+++ torch/csrc/api/include/torch/nn/modules/dropout.h
@@ -11,7 +11,7 @@ namespace torch {
 namespace nn {
 
 /// Options for `Dropout` and `FeatureDropout`.
-struct DropoutOptions {
+struct TORCH_API DropoutOptions {
   /* implicit */ DropoutOptions(double rate = 0.5);
   /// The probability with which a particular component of the input is set to
   /// zero.
@@ -36,7 +36,7 @@ class DropoutImplBase : public torch::nn::Cloneable<Derived> {
 ///
 /// See https://pytorch.org/docs/stable/nn.html#torch.nn.Dropout to learn more
 /// about the exact semantics of this module.
-class DropoutImpl : public detail::DropoutImplBase<DropoutImpl> {
+class TORCH_API DropoutImpl : public detail::DropoutImplBase<DropoutImpl> {
  public:
   using detail::DropoutImplBase<DropoutImpl>::DropoutImplBase;
   /// During training, applies a noise mask to the input tensor.
@@ -53,7 +53,7 @@ class DropoutImpl : public detail::DropoutImplBase<DropoutImpl> {
 /// [Dropout3d](https://pytorch.org/docs/stable/nn.html#torch.nn.Dropout3d) for
 /// 3-D features. This `FeatureDropout` module can instead deal with both 2-D
 /// and 3-D features.
-class FeatureDropoutImpl : public detail::DropoutImplBase<FeatureDropoutImpl> {
+class TORCH_API FeatureDropoutImpl : public detail::DropoutImplBase<FeatureDropoutImpl> {
  public:
   using detail::DropoutImplBase<FeatureDropoutImpl>::DropoutImplBase;
   /// During training, applies a noise mask to the input tensor.
