+++ torch/csrc/api/include/torch/optim/adam.h
@@ -18,7 +18,7 @@ class InputArchive;
 namespace torch {
 namespace optim {
 
-struct AdamOptions {
+struct TORCH_API AdamOptions {
   /* implicit */ AdamOptions(double learning_rate);
   TORCH_ARG(double, learning_rate);
   TORCH_ARG(double, beta1) = 0.9;
@@ -28,7 +28,7 @@ struct AdamOptions {
   TORCH_ARG(bool, amsgrad) = false;
 };
 
-class Adam : public Optimizer {
+class TORCH_API Adam : public Optimizer {
  public:
   template <typename ParameterContainer>
   explicit Adam(ParameterContainer&& parameters, const AdamOptions& options)
