+++ torch/csrc/api/include/torch/utils.h
@@ -9,10 +9,10 @@ using autograd::AutoGradMode;
 
 // A RAII, thread local (!) guard that stops future operations from building
 // gradients.
-struct NoGradGuard : public AutoGradMode {
+struct TORCH_API NoGradGuard : public AutoGradMode {
   NoGradGuard() : AutoGradMode(/*enabled=*/false) {}
 };
 
 /// Sets the global random seed for all newly created CPU and CUDA tensors.
-void manual_seed(uint64_t seed);
+void TORCH_API manual_seed(uint64_t seed);
 } // namespace torch
